server-py/requirements.txt
uvicorn
fastapi
python-multipart
numpy<2.0
tqdm
motor
pydantic
pydantic[email]
python-jose
pymongo[srv]
passlib
bcrypt
librosa
pydub
httpx
scikit-learn==1.2.2
authlib
itsdangerous
python-dotenv
starlette